+++ src/mpl_stereo/example_data.py
@@ -22,3 +22,7 @@ def church_left_right():
     church_left_data = mpimg.imread(MPL_STEREO_DATA_DIR / 'church_left.jpg')
     church_right_data = mpimg.imread(MPL_STEREO_DATA_DIR / 'church_right.jpg')
     return church_left_data, church_right_data
+
+def church_left_cropped():
+    church_left_data = mpimg.imread(MPL_STEREO_DATA_DIR / 'church_left_cropped.jpg')
+    return church_left_data